% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflected_density_fcn.R
\name{density.reflected}
\alias{density.reflected}
\title{Modification of GoFKernel density.reflected function}
\usage{
\method{density}{reflected}(x, lower = -Inf, upper = Inf, weights = NULL, ...)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{lower}{The lower bound across which data will be reflected.}

\item{upper}{The upper bound across which data will be reflected.}

\item{weights}{An optional vector of weights with length of \code{x}.}
}
\value{
An S3 density.reflected class object?
}
\description{
Modification of GoFKernel density.reflected function
}
\keyword{internal}
