% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_fcns.R
\name{dscaler}
\alias{dscaler}
\title{Rescale a PDF over an interval}
\usage{
dscaler(pdf, a, b)
}
\arguments{
\item{pdf}{A univariate probability density distribution.}

\item{a, b}{The new lower and upper limits of the distribution.
The density will be rescaled such that the integral over the
interval \eqn{[a,b]} sums to unity.}
}
\value{
An S3 density object.
}
\description{
Rescale a PDF over an interval
}
\keyword{internal}
