% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_fcns.R
\name{hell}
\alias{hell}
\title{Calculate Hellinger's distance}
\usage{
hell(d1, d2, extrap = TRUE)
}
\arguments{
\item{d1, d2}{A density distribution.}

\item{extrap}{Logical. If axes limits of `d1` and `d2` differ, should
they be extrapolated to match? The extended region will have density of 0.}
}
\value{
A numeric value between 0 and 1 inclusive.
}
\description{
Calculate Hellinger's distance
}
