% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_fcns.R
\name{schroenr}
\alias{schroenr}
\title{Calculate Schoener's D metric of niche overlap}
\usage{
schroenr(d1, d2)
}
\arguments{
\item{d1}{A density distribution.}

\item{d2}{A density distribution.}
}
\value{
A numeric value between 0 and 1 inclusive.
}
\description{
Calculate Schoener's D metric of niche overlap
}
