% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_density_fcn.R
\name{strappy}
\alias{strappy}
\title{Borrajo et al. 2017 bandwidth selection}
\usage{
strappy(Y, w, method = "brt")
}
\arguments{
\item{Y}{A numeric vector of observations.}

\item{w}{A function that gives the probability of observation
# at any single value in the range of \code{Y}.}

\item{method}{One of \code{c('rt', 'brt')}.
\code{rt} is the rule-of-thumb estimate; \code{brt} is a bootstrap estimate
with the rule-of-thumb as the pilot.}
}
\description{
Borrajo et al. 2017 bandwidth selection
}
\keyword{internal}
