% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformed_density_fcn.R
\name{transdens}
\alias{transdens}
\title{Transform a Biased Sample to Estimate Density}
\usage{
transdens(x, w, reflect = FALSE, a = NULL, b = NULL, ...)
}
\arguments{
\item{x}{A numeric vector from which the estimate is to be computed.}

\item{w}{A function that gives the probability of observation
at any single value in the range of \code{x}.}

\item{reflect}{Logical: should boundary reflection be applied?}

\item{a}{The lower limit for density estimation,
on the original, untransformed scale. Default is \code{min(x)}.}

\item{b}{The upper limit for density estimation,
on the original, untransformed scale. Default is \code{max(x)}.}

\item{...}{Further arguments passed on to \code{base::density}.}
}
\description{
Transform a Biased Sample to Estimate Density
}
