% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_fcns.R
\name{schoenr}
\alias{schoenr}
\title{Calculate Schoener's D metric of niche overlap}
\usage{
schoenr(d1, d2)
}
\arguments{
\item{d1}{A density distribution.}

\item{d2}{A density distribution.}
}
\value{
A numeric value between 0 and 1 inclusive.
}
\description{
This is a continuous version of the method introduced by Schoener 1968.
}
\references{
\insertRef{Schoener68}{kerneval}
}
