% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_fcns.R
\name{schoenr}
\alias{schoenr}
\title{Calculate Schoener's D metric of niche overlap}
\usage{
schoenr(d1, d2, a = NULL, b = NULL)
}
\arguments{
\item{d1}{A density distribution.}

\item{d2}{A density distribution.}

\item{a}{The lower and upper limits of comparison.
Each density will be rescaled such that the integral over the
interval \code{[a,b]} sums to unity. If empty, the minimum and maximum
of the intersection of \code{d1$x} and \code{d2$x} will be used.}

\item{b}{The lower and upper limits of comparison.
Each density will be rescaled such that the integral over the
interval \code{[a,b]} sums to unity. If empty, the minimum and maximum
of the intersection of \code{d1$x} and \code{d2$x} will be used.}
}
\value{
A numeric value between 0 and 1 inclusive.
}
\description{
Schoener's D metric quantifies niche overlap between two discretised
probability functions. The \code{schoenr} function modifies the metric for
continuous probability distributions.
}
\details{
D was originally defned as the sum of the absolute difference in resource use
frequencies for every category i. The sum is rescaled by 0.5 to give the
per-niche value, and then subtracted from 1 so as to be a similarity metric
(i.e. D = 1 indicates identical niches). \code{schoenr} replaces the sum
with an integral on the absolute difference between two estimated
kernel density functions.
}
\references{
\insertRef{Schoener68}{kerneval}
}
