% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformed_density_fcn.R
\name{transdens}
\alias{transdens}
\title{Transform a biased sample to estimate probability density}
\usage{
transdens(x, w, reflect = FALSE, a = NULL, b = NULL, ...)
}
\arguments{
\item{x}{A numeric vector from which the estimate is to be computed.}

\item{w}{A function that gives the probability of observation
at any single value in the range of \code{x}.}

\item{reflect}{Logical: should boundary reflection be applied?}

\item{a}{The lower limit for density estimation,
on the original, untransformed scale. Default is \code{min(x)}.}

\item{b}{The upper limit for density estimation,
on the original, untransformed scale. Default is \code{max(x)}.}

\item{...}{Further arguments passed on to \code{\link[stats]{density}}.}
}
\value{
An S3 density.
}
\description{
Calculate a kernel density estimate while correcting for selection bias
by transforming the data.
}
\details{
\code{transdens} implements the strategy of Barmi and Simonoff (2000)
to correct for selection bias in kernel density estimation. The method
(1) transforms the empirical data based on the cumulative distribution function
of the bias function \code{w}, (2) scales the density so it integrates to
unity, and then (3) back-transforms the density to the original scale.

Depending on the shape of the true probability distribution function and
the bias function, analysts would be wise to inspect kernel density plots on
the transformed scale, just as one might plot estimates (on the original
scale) when selecting a bandwidth. In particular, one should consider whether
the transformed distribution has a long tail or otherwise is difficult
to estimate. If the density estimation problem seems more straightforward
on the original scale, one could weight the kernel density estimate
with \code{wdens()} instead of transforming the data.
}
\references{
\insertRef{Barmi00}{kerneval}
}
\seealso{
\code{\link{wdens}}
}
