% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_density_fcn.R
\name{wdens}
\alias{wdens}
\title{Weight a Biased Sample to Estimate Density}
\usage{
wdens(x, w, bw = "brt", reflect = FALSE, a = NULL, b = NULL, ...)
}
\arguments{
\item{x}{A numeric vector from which the estimate is to be computed.}

\item{w}{A function that gives the probability of observation
at any single value in the range of \code{x}.}

\item{bw}{A method to estimate the kernel bandwidth from Borrajo et al. 2017.
\code{rt} is the rule-of-thumb estimate; \code{brt} is a bootstrap estimate
with the rule-of-thumb as the pilot.}

\item{reflect}{Logical: should boundary reflection be applied?}

\item{a}{The lower limit for density estimation,
on the original, untransformed scale. Default is \code{min(x)}.}

\item{b}{The upper limit for density estimation,
on the original, untransformed scale. Default is \code{max(x)}.}

\item{...}{Further arguments passed on to \code{base::density}.}
}
\description{
Weight a Biased Sample to Estimate Density
}
